requirements.txt
alembic==1.12.0
blinker==1.6.2
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
cryptography==41.0.3
firebase-admin==6.2.0
Flask==2.3.2
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WhooshAlchemy==0.56
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==2.0.2
grpcio==1.59.2
grpcio-status==1.59.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
Levenshtein==0.23.0
Mako==1.2.4
MarkupSafe==2.1.3
msgpack==1.0.7
mysql-connector-python==8.3.0
numpy==1.26.0
pandas==2.1.1
Pillow==10.0.1
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.0.3
pyparsing==3.1.1
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-Levenshtein==0.23.0
pytz==2023.3
rapidfuzz==3.5.2
requests==2.31.0
rsa==4.9
six==1.16.0
sms-python-client==1.0.4
SQLAlchemy==2.0.21
SQLAlchemy-serializer==1.4.1
stripe==7.10.0
tendo==0.3.0
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
Werkzeug==2.3.6
Whoosh==2.7.4